updating stars to integer in model/migration

+++ backend/db/models/review.js
@@ -55,7 +55,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     stars: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.INTEGER,
       allowNull: false,
       validate: {
         min: 1,
